Move `tweet_text_generator` to utils
