+++ bot/utils/tweet_text_generator.py
